% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbly.R
\name{bubbly}
\alias{bubbly}
\title{Add beautiful interactive bubbles within Shing and RMarkdown backgrounds}
\usage{
bubbly(theme = "ocean")
}
\arguments{
\item{theme}{Name of the bubbles theme. See the vignette for a list of themes.}
}
\value{
Themed interactive bubbles for Shiny and RMarkdown backgrounds
}
\description{
Add beautiful interactive bubbles within Shing and RMarkdown backgrounds
}
\examples{
 if (interactive()) {

 ui <- fluidPage(

 bubbly(theme = "meteor")

 )


 server <- function(input, output) {}



 shinyApp(ui = ui, server = server)




}


}
